# backend/app.py
from flask import Flask, request, jsonify
from flask_cors import CORS
from datetime import datetime
import os

app = Flask(__name__)
CORS(app)

# Health check endpoint
@app.route('/api/health', methods=['GET'])
def health_check():
    return jsonify({
        'status': 'healthy',
        'timestamp': datetime.utcnow().isoformat(),
        'service': 'scout-analytics-api'
    })

# Transactions endpoint
@app.route('/api/transactions', methods=['GET'])
def get_transactions():
    # TODO: Connect to actual database
    # For now, return mock data structure
    page = request.args.get('page', 1, type=int)
    per_page = request.args.get('per_page', 50, type=int)
    
    return jsonify({
        'data': [],
        'pagination': {
            'page': page,
            'per_page': per_page,
            'total': 0,
            'pages': 0
        }
    })

# Products endpoint
@app.route('/api/products', methods=['GET'])
def get_products():
    category = request.args.get('category')
    search = request.args.get('search')
    
    return jsonify({
        'data': [],
        'total': 0
    })

# Regions endpoint
@app.route('/api/regions', methods=['GET'])
def get_regions():
    return jsonify({
        'data': [],
        'total': 0
    })

# Analytics summary endpoint
@app.route('/api/analytics/summary', methods=['GET'])
def get_analytics_summary():
    date_from = request.args.get('date_from')
    date_to = request.args.get('date_to')
    region = request.args.get('region')
    
    return jsonify({
        'total_revenue': 0,
        'total_transactions': 0,
        'average_basket_size': 0,
        'top_products': [],
        'revenue_by_region': []
    })

# Brand performance endpoint
@app.route('/api/analytics/brand-performance', methods=['GET'])
def get_brand_performance():
    brand = request.args.get('brand')
    date_from = request.args.get('date_from')
    date_to = request.args.get('date_to')
    
    return jsonify({
        'brand': brand,
        'metrics': {
            'revenue': 0,
            'units_sold': 0,
            'market_share': 0,
            'growth_rate': 0
        },
        'trends': []
    })

# Consumer insights endpoint
@app.route('/api/analytics/consumer-insights', methods=['GET'])
def get_consumer_insights():
    segment = request.args.get('segment')
    
    return jsonify({
        'demographics': {},
        'purchase_patterns': [],
        'preferences': []
    })

if __name__ == '__main__':
    port = int(os.environ.get('PORT', 5000))
    app.run(host='0.0.0.0', port=port, debug=True)